test(providers): tighten orphan lifecycle guard
