+++ src/core/providers/registry/lifecycle.rs
@@ -305,9 +305,19 @@ pub static PROVIDER_ORPHAN_BASELINE: &[ProviderOrphanBaselineEntry] = &[
         "macro-generated chat provider awaiting GH837 disposition approval",
     ),
     baseline(
+        "deepgram",
+        "non-llm-lane",
+        "audio provider module without a gateway LLM factory path",
+    ),
+    baseline(
         "deepl",
         "non-llm-lane",
         "translation provider uses macro-generated LLMProvider surface",
+    ),
+    baseline(
+        "elevenlabs",
+        "non-llm-lane",
+        "audio provider module without a gateway LLM factory path",
     ),
     baseline(
         "empower",
@@ -486,332 +496,5 @@ const fn entry(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::collections::BTreeSet;
-    use std::fs;
-    use std::path::{Path, PathBuf};
-
-    const NATIVE_RUNTIME_MODULES: &[&str] = &[
-        "anthropic",
-        "azure",
-        "azure_ai",
-        "bedrock",
-        "cloudflare",
-        "cohere",
-        "fal_ai",
-        "gemini",
-        "github_copilot",
-        "mistral",
-        "openai",
-        "openai_like",
-        "replicate",
-        "vertex_ai",
-    ];
-
-    const PROVIDER_IMPL_MARKERS: &[&str] = &[
-        "impl LLMProvider for",
-        "define_http_provider_with_hooks!(",
-        "define_pooled_http_provider_with_hooks!(",
-        "define_openai_compatible_provider!(",
-    ];
-
-    fn lifecycle_module_names() -> BTreeSet<&'static str> {
-        PROVIDER_MODULE_LIFECYCLE
-            .iter()
-            .map(|entry| entry.module_name)
-            .collect()
-    }
-
-    fn orphan_baseline_module_names() -> BTreeSet<&'static str> {
-        PROVIDER_ORPHAN_BASELINE
-            .iter()
-            .map(|entry| entry.module_name)
-            .collect()
-    }
-
-    fn lifecycle_for(module_name: &str) -> ProviderModuleLifecycle {
-        PROVIDER_MODULE_LIFECYCLE
-            .iter()
-            .find(|entry| entry.module_name == module_name)
-            .unwrap_or_else(|| panic!("missing lifecycle entry for {module_name}"))
-            .lifecycle
-    }
-
-    fn providers_dir() -> PathBuf {
-        Path::new(env!("CARGO_MANIFEST_DIR")).join("src/core/providers")
-    }
-
-    fn provider_directories() -> BTreeSet<String> {
-        fs::read_dir(providers_dir())
-            .expect("providers directory should be readable")
-            .filter_map(|entry| {
-                let entry = entry.expect("provider directory entry should be readable");
-                if !entry
-                    .file_type()
-                    .expect("file type should be readable")
-                    .is_dir()
-                {
-                    return None;
-                }
-                entry.file_name().into_string().ok()
-            })
-            .collect()
-    }
-
-    fn is_native_runtime_module(module_name: &str) -> bool {
-        NATIVE_RUNTIME_MODULES.contains(&module_name)
-    }
-
-    fn directory_contains_provider_impl_marker(module_name: &str) -> bool {
-        if matches!(module_name, "macros" | "registry") {
-            return false;
-        }
-
-        let mut pending_dirs = vec![providers_dir().join(module_name)];
-        while let Some(dir) = pending_dirs.pop() {
-            for entry in fs::read_dir(&dir).unwrap_or_else(|err| {
-                panic!("provider directory {dir:?} should be readable: {err}")
-            }) {
-                let entry = entry.expect("provider directory entry should be readable");
-                let path = entry.path();
-                let file_type = entry.file_type().expect("file type should be readable");
-
-                if file_type.is_dir() {
-                    pending_dirs.push(path);
-                    continue;
-                }
-
-                if path.extension().is_some_and(|extension| extension == "rs") {
-                    let source = fs::read_to_string(&path).unwrap_or_else(|err| {
-                        panic!("provider source {path:?} should be readable: {err}")
-                    });
-                    if PROVIDER_IMPL_MARKERS
-                        .iter()
-                        .any(|marker| source.contains(marker))
-                    {
-                        return true;
-                    }
-                }
-            }
-        }
-
-        false
-    }
-
-    #[test]
-    fn lifecycle_classifies_phase0_key_provider_modules() {
-        assert_eq!(lifecycle_for("bedrock"), ProviderModuleLifecycle::Wire);
-        assert_eq!(
-            lifecycle_for("vertex_ai"),
-            if cfg!(feature = "providers-extra") {
-                ProviderModuleLifecycle::Wire
-            } else {
-                ProviderModuleLifecycle::Stub
-            }
-        );
-        assert_eq!(
-            lifecycle_for("azure"),
-            if cfg!(feature = "providers-extra") {
-                ProviderModuleLifecycle::Wire
-            } else {
-                ProviderModuleLifecycle::Stub
-            }
-        );
-        assert_eq!(
-            lifecycle_for("azure_ai"),
-            if cfg!(feature = "providers-extra") {
-                ProviderModuleLifecycle::Wire
-            } else {
-                ProviderModuleLifecycle::Stub
-            }
-        );
-        assert_eq!(
-            lifecycle_for("github_copilot"),
-            if cfg!(feature = "providers-extended") {
-                ProviderModuleLifecycle::Wire
-            } else {
-                ProviderModuleLifecycle::Stub
-            }
-        );
-        assert_eq!(
-            lifecycle_for("cohere"),
-            if cfg!(feature = "providers-extended") {
-                ProviderModuleLifecycle::Wire
-            } else {
-                ProviderModuleLifecycle::Stub
-            }
-        );
-        assert_eq!(
-            lifecycle_for("fal_ai"),
-            if cfg!(feature = "providers-extended") {
-                ProviderModuleLifecycle::Wire
-            } else {
-                ProviderModuleLifecycle::Stub
-            }
-        );
-        assert_eq!(
-            lifecycle_for("replicate"),
-            if cfg!(feature = "providers-extended") {
-                ProviderModuleLifecycle::Wire
-            } else {
-                ProviderModuleLifecycle::Stub
-            }
-        );
-        assert_eq!(
-            lifecycle_for("gemini"),
-            if cfg!(feature = "providers-extended") {
-                ProviderModuleLifecycle::Wire
-            } else {
-                ProviderModuleLifecycle::Stub
-            }
-        );
-    }
-
-    #[test]
-    fn lifecycle_wire_entries_are_native_runtime_modules() {
-        let actual = PROVIDER_MODULE_LIFECYCLE
-            .iter()
-            .filter(|entry| entry.lifecycle == ProviderModuleLifecycle::Wire)
-            .map(|entry| entry.module_name)
-            .collect::<BTreeSet<_>>();
-        let expected = [
-            "anthropic",
-            #[cfg(feature = "providers-extra")]
-            "azure",
-            #[cfg(feature = "providers-extra")]
-            "azure_ai",
-            "bedrock",
-            "cloudflare",
-            #[cfg(feature = "providers-extended")]
-            "cohere",
-            #[cfg(feature = "providers-extended")]
-            "fal_ai",
-            #[cfg(feature = "providers-extended")]
-            "gemini",
-            #[cfg(feature = "providers-extended")]
-            "github_copilot",
-            #[cfg(feature = "providers-extended")]
-            "replicate",
-            "mistral",
-            "openai",
-            "openai_like",
-            #[cfg(feature = "providers-extra")]
-            "vertex_ai",
-        ]
-        .into_iter()
-        .collect::<BTreeSet<_>>();
-
-        assert_eq!(actual, expected);
-    }
-
-    #[test]
-    fn lifecycle_covers_every_provider_directory() {
-        let actual = provider_directories();
-        let declared = lifecycle_module_names()
-            .into_iter()
-            .map(str::to_string)
-            .collect::<BTreeSet<_>>();
-
-        assert_eq!(actual, declared);
-    }
-
-    #[test]
-    fn lifecycle_has_no_delete_decisions_without_owner_confirmation() {
-        assert!(
-            PROVIDER_MODULE_LIFECYCLE
-                .iter()
-                .all(|entry| entry.lifecycle != ProviderModuleLifecycle::Delete),
-            "Delete lifecycle requires explicit owner confirmation"
-        );
-    }
-
-    #[test]
-    fn lifecycle_entries_have_reasons() {
-        for entry in PROVIDER_MODULE_LIFECYCLE {
-            assert!(
-                !entry.reason.trim().is_empty(),
-                "{} lifecycle entry must include a reason",
-                entry.module_name
-            );
-        }
-    }
-
-    #[test]
-    fn lifecycle_blocks_unapproved_provider_impl_modules() {
-        let baseline = orphan_baseline_module_names();
-        let mut unapproved = Vec::new();
-
-        for module_name in provider_directories() {
-            if !directory_contains_provider_impl_marker(&module_name) {
-                continue;
-            }
-            if is_native_runtime_module(&module_name) {
-                continue;
-            }
-            if baseline.contains(module_name.as_str()) {
-                continue;
-            }
-            unapproved.push(module_name);
-        }
-
-        assert!(
-            unapproved.is_empty(),
-            "unapproved provider implementation modules must be wired, deleted, demoted, or added to the GH837 baseline: {unapproved:?}"
-        );
-    }
-
-    #[test]
-    fn orphan_baseline_entries_are_live_and_bounded() {
-        let provider_dirs = provider_directories();
-        let mut seen = BTreeSet::new();
-
-        for entry in PROVIDER_ORPHAN_BASELINE {
-            assert!(
-                seen.insert(entry.module_name),
-                "{} appears more than once in the orphan baseline",
-                entry.module_name
-            );
-            assert!(
-                provider_dirs.contains(entry.module_name),
-                "{} baseline entry must reference an existing provider directory",
-                entry.module_name
-            );
-            assert!(
-                directory_contains_provider_impl_marker(entry.module_name),
-                "{} baseline entry must still contain a provider implementation marker",
-                entry.module_name
-            );
-            assert!(
-                !is_native_runtime_module(entry.module_name),
-                "{} is natively reachable and should not be in the orphan baseline",
-                entry.module_name
-            );
-            assert!(
-                matches!(
-                    entry.lane,
-                    "delete-native" | "demote-to-catalog" | "non-llm-lane" | "exempt"
-                ),
-                "{} baseline entry has unsupported lane {}",
-                entry.module_name,
-                entry.lane
-            );
-            assert_eq!(entry.issue, "GH837");
-            assert!(
-                !entry.owner.trim().is_empty(),
-                "{} baseline entry must include an owner",
-                entry.module_name
-            );
-            assert!(
-                !entry.expires.trim().is_empty(),
-                "{} baseline entry must include an expiry condition",
-                entry.module_name
-            );
-            assert!(
-                !entry.reason.trim().is_empty(),
-                "{} baseline entry must include a reason",
-                entry.module_name
-            );
-        }
-    }
-}
+#[path = "lifecycle_tests.rs"]
+mod tests;
